Change firebase version to 19.7.0, rename tests packages, patch HomeScreenActivityToUserProfileActivity test
