+++ app/build.gradle
@@ -72,7 +72,7 @@ dependencies {
     implementation "androidx.fragment:fragment:1.3.2"
     implementation 'androidx.room:room-runtime:2.2.6'
 
-    implementation 'com.google.firebase:firebase-database:19.5.1'
+    implementation 'com.google.firebase:firebase-database:19.7.0'
 
 
     testImplementation 'junit:junit:4.13.2'
