+++ app/src/androidTest/java/com/sdp/movemeet/view/chat/MessageHolderTest.java
@@ -1,0 +1,5 @@
+package com.sdp.movemeet.view.chat;
+
+public class MessageHolderTest {
+
+}
