+++ app/src/androidTest/java/com/sdp/movemeet/view/chat/MessageViewHolderTest.java
@@ -1,0 +1,7 @@
+package com.sdp.movemeet.view.chat;
+
+public class MessageViewHolderTest {
+    private String DUMMY_TAG = "MessageViewHolder";
+
+
+}
