+++ app/src/androidTest/java/com/sdp/movemeet/view/map/MiniMapFragmentTest.java
