+++ app/src/androidTest/java/com/sdp/movemeet/view/navigation/ProfileActivityNavigationTest.java
@@ -1,0 +1,113 @@
+package com.sdp.movemeet.view.navigation;
+
+import androidx.test.ext.junit.rules.ActivityScenarioRule;
+
+import com.sdp.movemeet.view.home.HomeScreenActivity;
+
+import org.junit.Rule;
+import org.junit.Test;
+
+import static androidx.test.espresso.action.ViewActions.click;
+import static androidx.test.espresso.contrib.DrawerMatchers.isClosed;
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
+import static androidx.test.espresso.matcher.ViewMatchers.withText;
+import static org.hamcrest.Matchers.allOf;
+
+public class ProfileActivityNavigationTest {
+    @Rule
+    public ActivityScenarioRule<HomeScreenActivity> testRule = new ActivityScenarioRule<>(HomeScreenActivity.class);
+
+    @Test
+    public void emptyTest(){}
+    /*@Before
+    public void signIn(){
+        onView(withId(R.id.signInButton)).perform(click());
+        onView(withId(R.id.edit_text_email)).perform(replaceText("antho2@gmail.com"), closeSoftKeyboard());
+        onView(withId(R.id.edit_text_password)).perform(replaceText("234567"), closeSoftKeyboard());
+        onView(withId(R.id.button_login)).perform(click());
+        try {
+            Thread.sleep(2500);
+        } catch (InterruptedException e) {
+            assert (false);
+        }
+        // Open Drawer to click on navigation.
+        onView(withId(R.id.drawer_layout)).check(matches(isClosed(Gravity.LEFT))).perform(DrawerActions.open());
+        try{
+            Thread.sleep(500);
+        }catch(Exception e){}
+
+        onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_edit_profile));
+
+        try{
+            Thread.sleep(500);
+        }catch(Exception e){}
+
+        // Open Drawer to click on navigation.
+        onView(withId(R.id.drawer_layout)).check(matches(isClosed(Gravity.LEFT))).perform(DrawerActions.open());
+        try{
+            Thread.sleep(2500);
+        }catch(Exception e){}
+    }
+
+    @Test
+    public void ProfileActivityToProfileActivity() {
+        onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_edit_profile));
+        onView(withId(R.id.button_update_profile)).check(matches(isDisplayed()));
+        logout();
+    }
+
+    @Test
+    public void ProfileActivityToActivityUpload() {
+        onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_add_activity));
+        onView(withId(R.id.buttonConfirmUpload)).check(matches(isDisplayed()));
+        logout();
+    }
+
+    @Test
+    public void UploadActivityToStartActivity() {
+        onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_start_activity));
+        onView(withId(R.id.activityChatDescription)).check(matches(isDisplayed()));
+        logout();
+    }
+
+    @Test
+    public void UploadActivityGotoHome() {
+        onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_home));
+        onView(withId(R.id.textView)).check(matches(isDisplayed()));
+        onView(allOf(withId(R.id.textView), withText("MOVEMEƎT")));
+        logout();
+    }
+
+    @Test
+    public void ProfileActivityGotoChat() {
+        onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_chat));
+        onView(withId(R.id.addMessageImageView)).check(matches(isDisplayed()));
+        logout();
+    }
+
+    @Test
+    public void ProfileActivity_logout() {
+        onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_logout));
+        onView(withId(R.id.button_login)).check(matches(isDisplayed()));
+    }
+
+    public void logout() {
+
+        try {
+            Thread.sleep(2000);
+        } catch (InterruptedException e) {
+            assert (false);
+        }
+
+        // Open Drawer to click on navigation.
+        onView(withId(R.id.drawer_layout))
+                .check(matches(isClosed(Gravity.LEFT))) // Left Drawer should be closed.
+                .perform(DrawerActions.open()); // Open Drawer
+
+        try{
+            Thread.sleep(500);
+        }catch(Exception e){}
+
+        onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_logout));
+    }*/
+}
