+++ app/src/androidTest/java/com/sdp/movemeet/view/navigation/securityTest/ChatActivitySecurityTest.java
@@ -1,0 +1,25 @@
+package com.sdp.movemeet.view.navigation.securityTest;
+
+import androidx.test.ext.junit.rules.ActivityScenarioRule;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+
+import com.sdp.movemeet.view.chat.ChatActivity;
+
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
+
+@RunWith(AndroidJUnit4.class)
+
+public class ChatActivitySecurityTest {
+
+    @Rule
+    public ActivityScenarioRule<ChatActivity> testRule = new ActivityScenarioRule<>(ChatActivity.class);
+
+    @Test
+    public void redirectionTest() {
+        //onView(withId(R.id.button_login)).check(matches(isDisplayed()));
+    }
+}
