+++ app/src/androidTest/java/com/sdp/movemeet/view/navigation/securityTest/ProfileActivitySecurityTest.java
@@ -1,0 +1,25 @@
+package com.sdp.movemeet.view.navigation.securityTest;
+
+import androidx.test.ext.junit.rules.ActivityScenarioRule;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+
+import com.sdp.movemeet.view.profile.ProfileActivity;
+
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
+
+@RunWith(AndroidJUnit4.class)
+
+public class ProfileActivitySecurityTest {
+
+    @Rule
+    public ActivityScenarioRule<ProfileActivity> testRule = new ActivityScenarioRule<>(ProfileActivity.class);
+
+    @Test
+    public void redirectionTest() {
+        //onView(withId(R.id.button_login)).check(matches(isDisplayed()));
+    }
+}
