+++ app/src/androidTest/java/com/sdp/movemeet/view/profile/HomeScreenActivityToUserProfileActivity.java
@@ -1,0 +1,93 @@
+package com.sdp.movemeet.view.profile;
+
+
+import android.view.Gravity;
+
+import androidx.test.espresso.contrib.DrawerActions;
+import androidx.test.espresso.contrib.NavigationViewActions;
+import androidx.test.filters.LargeTest;
+import androidx.test.rule.ActivityTestRule;
+import androidx.test.runner.AndroidJUnit4;
+
+import com.sdp.movemeet.view.home.HomeScreenActivity;
+import com.sdp.movemeet.R;
+
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.action.ViewActions.click;
+import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
+import static androidx.test.espresso.action.ViewActions.replaceText;
+import static androidx.test.espresso.assertion.ViewAssertions.matches;
+import static androidx.test.espresso.contrib.DrawerMatchers.isClosed;
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
+import static androidx.test.espresso.matcher.ViewMatchers.withText;
+import static org.hamcrest.Matchers.allOf;
+
+@LargeTest
+@RunWith(AndroidJUnit4.class)
+public class HomeScreenActivityToUserProfileActivity {
+
+    @Rule
+    public ActivityTestRule<HomeScreenActivity> mActivityTestRule = new ActivityTestRule<>(HomeScreenActivity.class);
+
+    @Test
+    public void homeScreenActivityToUserProfileActivity() {
+        onView(withId(R.id.signInButton)).perform(click());
+        onView(withId(R.id.edit_text_email)).perform(replaceText("antho2@gmail.com"), closeSoftKeyboard());
+        onView(withId(R.id.edit_text_password)).perform(replaceText("234567"), closeSoftKeyboard());
+        onView(withId(R.id.button_login)).perform(click());
+        try {
+            Thread.sleep(2000);
+        } catch (InterruptedException e) {
+            assert (false);
+        }
+
+        // Open Drawer to click on navigation.
+        onView(withId(R.id.drawer_layout))
+                .check(matches(isClosed(Gravity.LEFT))) // Left Drawer should be closed.
+                .perform(DrawerActions.open()); // Open Drawer
+
+        try{
+            Thread.sleep(500);
+        }catch(Exception e){}
+
+        // Start the screen of your activity.
+        onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_edit_profile));
+        try{
+            Thread.sleep(2000);
+        }catch(Exception e){}
+
+        onView(withId(R.id.text_view_activity_profile_name)).check(matches(withText("Anthony Guinchard")));
+
+        onView(withId(R.id.text_view_activity_profile_email)).check(matches(withText("antho2@gmail.com")));
+
+        onView(withId(R.id.text_view_activity_profile_phone)).check(matches(withText("000000000000")));
+
+        onView(withId(R.id.text_view_activity_profile_description)).check(matches(withText("")));
+        logout();
+
+    }
+
+    public void logout() {
+
+        try {
+            Thread.sleep(1000);
+        } catch (InterruptedException e) {
+            assert (false);
+        }
+
+        // Open Drawer to click on navigation.
+        onView(withId(R.id.drawer_layout))
+                .check(matches(isClosed(Gravity.LEFT))) // Left Drawer should be closed.
+                .perform(DrawerActions.open()); // Open Drawer
+
+        try{
+            Thread.sleep(500);
+        }catch(Exception e){}
+
+        onView(withId(R.id.nav_view)).perform(NavigationViewActions.navigateTo(R.id.nav_logout));
+    }
+}
