--- app/src/androidTest/java/com/sdp/movemeet/viewTest/chatTest/MessageHolderTest.java
@@ -1,5 +1,0 @@
-package com.sdp.movemeet.viewTest.chatTest;
-
-public class MessageHolderTest {
-
-}
