--- app/src/androidTest/java/com/sdp/movemeet/viewTest/chatTest/MessageViewHolderTest.java
@@ -1,7 +1,0 @@
-package com.sdp.movemeet.viewTest.chatTest;
-
-public class MessageViewHolderTest {
-    private String DUMMY_TAG = "MessageViewHolder";
-
-
-}
